Fix classes to standalone class files.
