+++ tests/test_app/TestApp/Controller/Component/BlueberryComponent.php
@@ -1,0 +1,26 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Controller\Component;
+
+use Cake\Controller\Component;
+
+class BlueberryComponent extends Component
+{
+    /**
+     * testName property
+     *
+     * @var string|null
+     */
+    public $testName;
+
+    /**
+     * initialize method
+     *
+     * @param array $config
+     * @return void
+     */
+    public function initialize(array $config): void
+    {
+        $this->testName = 'BlueberryComponent';
+    }
+}
