+++ tests/test_app/TestApp/Error/Exception/MissingWidgetThing.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Error\Exception;
+
+use Exception;
+
+/**
+ * Exception class for testing app error handlers and custom errors.
+ */
+class MissingWidgetThing extends Exception
+{
+}
