+++ tests/test_app/TestApp/Error/Exception/MissingWidgetThingException.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Error\Exception;
+
+use Cake\Http\Exception\NotFoundException;
+
+/**
+ * Exception class for testing app error handlers and custom errors.
+ */
+class MissingWidgetThingException extends NotFoundException
+{
+}
