+++ tests/test_app/TestApp/Error/Thing/DebuggableThing.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Error\Thing;
+
+class DebuggableThing
+{
+    public function __debugInfo()
+    {
+        return ['foo' => 'bar', 'inner' => new self()];
+    }
+}
