+++ tests/test_app/TestApp/Error/Thing/SecurityThing.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Error\Thing;
+
+class SecurityThing
+{
+    public $password = 'pass1234';
+}
